+++ app/Http/Controllers/AdminControllers/PostCategoryController.php
@@ -6,12 +6,14 @@ use App\Http\Controllers\Controller;
 use App\Models\PostCategory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Gate;
 
 class PostCategoryController extends Controller
 {
 
     public function index(Request $request)
     {
+        Gate::authorize('post_category.view');
         $query = PostCategory::query();
 
         // tìm kiếm theo tên danh mục
@@ -41,12 +43,13 @@ class PostCategoryController extends Controller
 
     public function create()
     {
+        Gate::authorize('post_category.create');
         return view('admin.post-categories.create');
     }
 
     public function store(Request $request)
     {
-
+        Gate::authorize('post_category.create');
         PostCategory::create([
             'name' => $request->name,
             'slug' => Str::slug($request->name),
@@ -61,6 +64,7 @@ class PostCategoryController extends Controller
 
     public function edit($id)
     {
+        Gate::authorize('post_category.update');
         $category = PostCategory::findOrFail($id);
 
         return view('admin.post-categories.edit', compact('category'));
@@ -68,7 +72,7 @@ class PostCategoryController extends Controller
 
     public function update(Request $request, $id)
     {
-
+        Gate::authorize('post_category.update');
         $category = PostCategory::findOrFail($id);
 
         $category->update([
@@ -84,7 +88,7 @@ class PostCategoryController extends Controller
 
     public function destroy($id)
     {
-
+        Gate::authorize('post_category.delete');
         $category = PostCategory::findOrFail($id);
 
         $category->delete();
